refactor: use one flowset for all sender flows of a given type
Multiple flows of the same type should be sent in a single FlowSet
in IPFIX. The previous code constructs a new FlowSet for each
individual spot, which is wasteful. This change partitions the
flows into withLocator/withoutLocator, and batches as many flows
as possible into a FlowSet.

Lower max UDP bytes to 1190, based on analysis of PSKReporter traffic.